refactor: made packages local instead of their own package
search, coin_product and utils have been made their own package with in
server, that can be locally imported
