--- server/coin_products/go.mod
@@ -1,28 +1,0 @@
-module github.com/shawnyu5/check_coin_product
-
-go 1.18
-
-replace github.com/shawnyu5/penguin-utils v0.0.0-00010101000000-000000000000 => ../utils
-
-require (
-	github.com/gocolly/colly v1.2.0
-	github.com/patrickmn/go-cache v2.1.0+incompatible
-	github.com/shawnyu5/penguin-utils v0.0.0-00010101000000-000000000000
-)
-
-require (
-	github.com/PuerkitoBio/goquery v1.8.0 // indirect
-	github.com/andybalholm/cascadia v1.3.1 // indirect
-	github.com/antchfx/htmlquery v1.2.4 // indirect
-	github.com/antchfx/xmlquery v1.3.10 // indirect
-	github.com/antchfx/xpath v1.2.0 // indirect
-	github.com/gobwas/glob v0.2.3 // indirect
-	github.com/golang/groupcache v0.0.0-20200121045136-8c9f03a8e57e // indirect
-	github.com/golang/protobuf v1.3.1 // indirect
-	github.com/kennygrant/sanitize v1.2.4 // indirect
-	github.com/saintfish/chardet v0.0.0-20120816061221-3af4cd4741ca // indirect
-	github.com/temoto/robotstxt v1.1.2 // indirect
-	golang.org/x/net v0.0.0-20220520000938-2e3eb7b945c2 // indirect
-	golang.org/x/text v0.3.7 // indirect
-	google.golang.org/appengine v1.6.7 // indirect
-)
